[Add]세션스토리지 url 받아오기 구현
