+++ src/App.jsx
@@ -1,6 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
 import GlobalStyle from './GrobalStyle';
-import Play from './components/Play/Play';
+import Play from './page/Play';
 import Record from './page/Record';
 import Header from './components/Header/Header';
