+++ src/components/Play/CustomPlay.css
@@ -3,9 +3,20 @@
   display: flex;
   flex-direction: column;
   line-height: 1;
-  width: 90%;
+  width: 85%;
   margin: auto;
   padding: 10px 15px;
   background-color: #fff;
   box-shadow: none;
+}
+
+.rhap_main-controls-button {
+  margin: 0 3px;
+  color: #00aac5;
+}
+
+.rhap_play-pause-button {
+  font-size: 70px;
+  width: 70px;
+  height: 70px;
 }
